+++ src/index.ts
@@ -1,5 +1,8 @@
+import { app } from './app';
+
 export type * from './types/AppConfigs.js';
 export type * from './types/AppOptions.js';
 export type * from './types/AppTypes.js';
 
 export { app } from './app';
+export default app;