+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
 ## [Unreleased]
-- 
+- No changes yet
+
+## 0.2.1 (2025-11-13)
+- Fixed few component and sidebar styles
+- removed `theme` setter property in `app` instance
+- changed `progress` property to function in `app` instance
+- changed name `locales` option to `localization` option in `AppConfig` type
 
 ## 0.2.0 (2025-11-12)
 - Initial library version release